Summarize a categorical variable by a grouping variable

cat_group_tbl() presents frequency counts and percentages (count, percent) for nominal or categorical variables by some grouping variable. Relative frequencies and percentages of each level of the primary categorical variable (row_var) within each level of the grouping variable (col_var) can be returned. Missing data can be excluded for either variable from the calculations. By default, the table is returned in the long format.

Usage

cat_group_tbl(
  data,
  row_var,
  col_var,
  na.rm.row_var = FALSE,
  na.rm.col_var = FALSE,
  only = NULL,
  ignore = NULL,
  pivot = "longer"
)

Arguments

data

A data frame.

row_var

A character string of the name of a column in data containing categorical data. This is the primary categorical variable. When pivoted to the wider format, the categories of this variable will appear in the rows of the table.

col_var

A character string of the name of a column in data containing categorical data. This is the primary categorical variable. When pivoted to the wider format, the categories of this variable will appear in the rows of the table.

na.rm.row_var

A logical value indicating whether missing values for row_var should be removed before calculations. Default is FALSE.

na.rm.col_var

A logical value indicating whether missing values for col_var should be removed before calculations. Default is FALSE.

only

A character string or vector of strings indicating the types of summary data to return. The default is NULL, which includes both counts and percentages. To return only one type, specify count or percent. Percentages are calculated column- wise, grouped by col_var.

ignore

A named character vector or list containing values to ignore from row_var and col_var.

pivot

A character string specifying the format of the returned summary table. The default is longer, which returns the data in long format. To return the data in wide format, use wider.

Value

A tibble displaying relative frequency counts and/or percentages of row_var, grouped by col_var. When the output is in wider format, columns prefixed with count_ and percent_ contain the frequency and proportion, respectively, for each distinct response value of row_var within each level of col_var.

Author

Ama Nyame-Mensah

Examples

cat_group_tbl(data = nlsy,
              row_var = "gender",
              col_var = "bthwht",
              pivot = "wider",
              only = "count")
#> # A tibble: 2 × 3
#>   gender count_bthwht_0 count_bthwht_1
#>    <dbl>          <int>          <int>
#> 1      0           1340            123
#> 2      1           1409            104

cat_group_tbl(data = nlsy,
              row_var = "birthord",
              col_var = "breastfed",
              pivot = "longer")
#> # A tibble: 16 × 4
#>    birthord breastfed count  percent
#>       <dbl>     <dbl> <int>    <dbl>
#>  1        1         0   431 0.286   
#>  2        2         0   573 0.380   
#>  3        3         0   319 0.212   
#>  4        4         0   115 0.0763  
#>  5        5         0    49 0.0325  
#>  6        6         0    13 0.00862 
#>  7        7         0     7 0.00464 
#>  8        8         0     1 0.000663
#>  9        1         1   614 0.418   
#> 10        2         1   499 0.340   
#> 11        3         1   242 0.165   
#> 12        4         1    77 0.0525  
#> 13        5         1    23 0.0157  
#> 14        6         1    11 0.00749 
#> 15        7         1     1 0.000681
#> 16        8         1     1 0.000681