Depressive Symptoms Data

Subset of data from the National Longitudinal Survey of Youth (NLSY) 1979 Children and Young Adults. This dataset includes survey responses about feelings and behaviors linked to depressive symptoms in children and young adults. For more information about the National Longitudinal Survey of Youth, visit: https://www.nlsinfo.org/.

Usage

depressive

Format

A data frame with 11,551 rows and 12 columns:

cid

Child identification number)

race

race of child (1 = Hispanic, 2 = Black, 3 = Non-Black,Non-Hispanic)

sex

sex of child (1 = male, 2 = female)

yob

year of child's bith

dep_1

how often child feels sad and blue (1 = often, 2 = sometimes, 3 = hardly ever)

dep_2

how often child feels nervous, tense, or on edge (1 = often, 2 = sometimes, 3 = hardly ever)

dep_3

how often child feels happy (1 = often, 2 = sometimes, 3 = hardly ever)

dep_4

how often child feels bored (1 = often, 2 = sometimes, 3 = hardly ever)

dep_5

how often child feels lonely (1 = often, 2 = sometimes, 3 = hardly ever)

dep_6

how often child feels tired or worn out (1 = often, 2 = sometimes, 3 = hardly ever)

dep_7

how often child feels excited about something (1 = often, 2 = sometimes, 3 = hardly ever)

dep_8

how often child feels too busy to get everything (1 = often, 2 = sometimes, 3 = hardly ever)