National Longitudinal Survey of Youth (NLSY) Data

These data are a subset from the National Longitudinal Survey of Youth (NLSY) 1979 Children and Young Adults.The data contains 2,976 observations and 10 variables.

For more information about the National Longitudinal Survey of Youth, visit https://www.nlsinfo.org/.

Usage

nlsy

Format

A tibble with 2,976 rows and 11 columns:

CID

Child identification number)

race

race of child (Hispanic, Black, Non-Black,Non-Hispanic)

gender

gender of child (1 = male, 0 = female)

birthord

birth order of child

magebirth

Age of mother at birth of child

bthwht

whether child was born low birth weight (1 = yes, 0 = no)

breastfed

whether child was breastfed (1 = yes, 0 = no)

medu

Highest grade completed by child’s mother

math

PIAT Math Standard Score

read

PIAT Reading Recognition Standard Score

hhnum

Number of household members in household