2020 Social Determinants of Health (SDOH) Data

Subset of data from the 2020 Social Determinants of Health (SDOH) Database. For more information about the 2020 SDOH Database, visit: https://www.ahrq.gov/sdoh/index.html.

sdoh

Format

A tibble with 3,229 rows and 29 columns:

YEAR

SDOH file year

COUNTYFIPS

State-county FIPS Code (5-digit)

STATEFIPS

State FIPS Code (2-digit)

STATE

State name

COUNTY

County name

REGION

Census region name

TERRITORY

Territory indicator (1= U.S. Territory, 0= U.S. State or DC)

ACS_PCT_AGE_0_4

Percentage of population between ages 0-4

ACS_PCT_AGE_5_9

Percentage of population between ages 5-9

ACS_PCT_AGE_10_14

Percentage of population between ages 10-14

ACS_PCT_AGE_15_17

Percentage of population between ages 15-17

NOAAC_PRECIPITATION_JAN

Monthly (January) precipitation (Inches)

NOAAC_PRECIPITATION_FEB

Monthly (February) precipitation (Inches)

NOAAC_PRECIPITATION_MAR

Monthly (March) precipitation (Inches)

NOAAC_PRECIPITATION_APR

Monthly (April) precipitation (Inches)

NOAAC_PRECIPITATION_MAY

Monthly (May) precipitation (Inches)

NOAAC_PRECIPITATION_JUN

Monthly (June) precipitation (Inches)

NOAAC_PRECIPITATION_JUL

Monthly (July) precipitation (Inches)

NOAAC_PRECIPITATION_AUG

Monthly (August) precipitation (Inches)

NOAAC_PRECIPITATION_SEP

Monthly (September) precipitation (Inches)

NOAAC_PRECIPITATION_OCT

Monthly (October) precipitation (Inches)

NOAAC_PRECIPITATION_NOV

Monthly (November) precipitation (Inches)

NOAAC_PRECIPITATION_DEC

Monthly (December) precipitation (Inches)

HHC_PCT_HHA_NURSING

Percentage of home health agencies offering nursing care services

HHC_PCT_HHA_PHYS_THERAPY

Percentage of home health agencies offering physical therapy services

HHC_PCT_HHA_OCC_THERAPY

Percentage of home health agencies offering occupational therapy services

HHC_PCT_HHA_SPEECH

Percentage of home health agencies offering speech pathology services

HHC_PCT_HHA_MEDICAL

Percentage of home health agencies offering medical social services

HHC_PCT_HHA_AIDE

Percentage of home health agencies offering home health aide services