Panel Study of Income Dynamics (PSID) Transition into Adulthood Supplement (TAS) Data

Subset of data from the Panel Study of Income Dynamics (PSID) Transition into Adulthood Supplement. This dataset includes information from young adults about how they spend their free time, including participation in organized activities such as clubs, sports or athletic teams, social-action groups, and other structured extracurricular engagements. For more information about the Panel Study of Income Dynamics, visit: https://psidonline.isr.umich.edu/GettingStarted.aspx.

tas

Format

A tibble with 2,526 rows and 8 columns:

pid

personal identification number)

sex

sex of individual (1 = female, 2 = male)

involved_arts

whether the individual participated in any organized activities related to art, music, or the theater in the last 12 months (1 = yes, 0 = no)

involved_sports

whether the individual was a member of any athletic or sports teams in the last 12 months (1 = yes, 0 = no)

involved_schoolClubs

whether the individual was involved with any high school or college clubs or student government in the last 12 months (1 = yes, 0 = no)

involved_election

whether the individual voted in the national election in November 2016 that was held to elect the President (1 = yes, 0 = no)

involved_socialActionGrps

whether the individual was involved in any political groups, solidarity or ethnic-support groups or social-action groups in the last 12 months (1 = yes, 0 = no)

involved_volunteer

whether the individual was involved in any unpaid volunteer or community service work in the last 12 months (1 = yes, 0 = no)