Summarize continuous variables by group

mean_group_tbl() presents descriptive statistics (mean, sd, minimum, maximum, number of non-missing observations) for interval (e.g., Test scores) and ratio level (e.g., Age) variables with the same variable stem by some grouping variable. A variable stem is a common prefix found in related variable names, often corresponding to similar survey items, that represents a shared concept before unique identifiers (like time points) are added. For example, in the stem_social_psych dataset, the two variables 'belong_belongStem_w1' and 'belong_belongStem_w2' share the variable stem 'belong_belongStem' (e.g., "I feel like an outsider in STEM"), with suffixes (_w1, _w2) indicating different measurement waves. By default, missing data are excluded from the calculations in a listwise fashion.

Usage

mean_group_tbl(
  data,
  var_stem,
  group,
  escape_stem = FALSE,
  ignore_stem_case = FALSE,
  group_type = "variable",
  group_name = NULL,
  escape_group = FALSE,
  ignore_group_case = FALSE,
  remove_group_non_alnum = TRUE,
  na_removal = "listwise",
  only = NULL,
  var_labels = NULL,
  ignore = NULL
)

Arguments

data

A data frame.

var_stem

A character string of a variable stem or the full name of a variable in data.

group

A character string of a variable in data or a pattern to use to search for variables in data.

escape_stem

A logical value indicating whether to escape var_stem. Default is FALSE.

ignore_stem_case

A logical value indicating whether the search for columns matching the supplied var_stem is case-insensitive. Default is FALSE.

group_type

A character string that defines the type of grouping variable. Should be one of pattern or variable. Default is variable, in which case the variable matching the group string will be searched for within data.

group_name

A character string piped to the final table to replace the name of group.

escape_group

A logical value indicating whether to escape string supplied to group.

ignore_group_case

A logical value indicating whether group is case-insensitive. Default is FALSE.

remove_group_non_alnum

A logical value indicating whether to remove all non- alphanumeric characters (anything that is not a letter or number) from group. Default is TRUE.

na_removal

A character string specifying how to remove missing values. Should be one of pairwise or listwise. Default is listwise.

only

A character string or vector of character strings of the types of summary data to return. Default is NULL, which returns mean (mean), standard deviation (sd), minimum value (min), maximum value (max), and non-missing responses (nobs).

var_labels

An optional named character vector or list where each element maps labels to variable names. If any element is unnamed or if any labels do not match variables in returned from data, all labels will be ignored and the table will be printed without them.

ignore

An optional named vector or list specifying values to exclude from the dataset and analysis. By default, NULL includes all available values. To omit values from variables returned by var_stem, use the provided stem as the name. To exclude values from both var_stem variables and a grouping variable in data, supply a list.

Value

A tibble presenting summary statistics (e.g., mean, standard deviation, minimum value, maximum, number of non-missing observations) for a set of variables sharing the same variable stem. The results are grouped by either a grouping variable in the data or by a pattern matched with variable names.

Author

Ama Nyame-Mensah

Examples

mean_group_tbl(data = stem_social_psych,
               var_stem = "belong_welcomedStem",
               group = "_w\\d",
               group_type = "pattern",
               na_removal = "pairwise",
               var_labels = c(belong_welcomedStem_w1 = "I feel welcomed in STEM workplaces",
                              belong_welcomedStem_w2 = "I feel welcomed in STEM workplaces"),
               group_name = "wave")
#> # A tibble: 2 × 8
#>   variable               variable_label      wave   mean    sd   min   max  nobs
#>   <chr>                  <chr>               <chr> <dbl> <dbl> <dbl> <dbl> <int>
#> 1 belong_welcomedStem_w1 I feel welcomed in… w1     3.73 1.01      1     5   706
#> 2 belong_welcomedStem_w2 I feel welcomed in… w2     3.93 0.957     1     5   302

mean_group_tbl(data = social_psy_data,
               var_stem = "belong",
               group = "gender",
               group_type = "variable",
               na_removal = "pairwise",
               var_labels = c(belong_1 = "I feel like I belong at this institution",
                              belong_2 = "I  feel like part of the community",
                              belong_3 = "I feel valued by this institution"),
               group_name = "gender_identity")
#> # A tibble: 18 × 8
#>    variable variable_label         gender_identity  mean    sd   min   max  nobs
#>    <chr>    <chr>                            <dbl> <dbl> <dbl> <dbl> <dbl> <int>
#>  1 belong_1 I feel like I belong …               1  3.39 1.14      1     5  3283
#>  2 belong_1 I feel like I belong …               2  4.09 0.977     1     5  5883
#>  3 belong_1 I feel like I belong …               3  3.38 1.27      1     5    48
#>  4 belong_1 I feel like I belong …               4  3.83 1.14      2     5    29
#>  5 belong_1 I feel like I belong …               5  3.80 1.12      1     5    54
#>  6 belong_1 I feel like I belong …               6  3.35 1.22      1     5    74
#>  7 belong_2 I  feel like part of …               1  2.97 1.16      1     5  3282
#>  8 belong_2 I  feel like part of …               2  2.24 1.17      1     5  5885
#>  9 belong_2 I  feel like part of …               3  3.04 1.38      1     5    48
#> 10 belong_2 I  feel like part of …               4  2.48 1.27      1     5    29
#> 11 belong_2 I  feel like part of …               5  2.73 1.33      1     5    55
#> 12 belong_2 I  feel like part of …               6  3.09 1.17      1     5    74
#> 13 belong_3 I feel valued by this…               1  3.46 1.00      1     5  3282
#> 14 belong_3 I feel valued by this…               2  3.97 0.927     1     5  5885
#> 15 belong_3 I feel valued by this…               3  3.35 1.000     1     5    48
#> 16 belong_3 I feel valued by this…               4  3.52 1.09      1     5    29
#> 17 belong_3 I feel valued by this…               5  3.56 1.05      1     5    55
#> 18 belong_3 I feel valued by this…               6  3    1.19      1     5    74

grouped_data <-
  data.frame(
    symptoms.t1 = sample(c(1:5, -999), replace = TRUE, size = 50),
    symptoms.t2 = sample(c(NA, 1:5, -999), replace = TRUE, size = 50)
  )

mean_group_tbl(data = grouped_data,
               var_stem = "symptoms",
               group = ".t\\d",
               group_type = "pattern",
               escape_group = TRUE,
               na_removal = "listwise",
               ignore = c(symptoms = -999))
#> # A tibble: 2 × 7
#>   variable    group  mean    sd   min   max  nobs
#>   <chr>       <chr> <dbl> <dbl> <dbl> <dbl> <int>
#> 1 symptoms.t1 t1     3.17  1.58     1     5    29
#> 2 symptoms.t2 t2     3     1.51     1     5    29