Summarize continuous variables

mean_tbl() presents descriptive statistics (mean, sd, minimum, maximum, number of non-missing observations) for interval (e.g., Test scores) and ratio level (e.g., Age) variables with the same variable stem. A variable stem is a common prefix found in related variable names, often corresponding to similar survey items, that represents a shared concept before unique identifiers (like timep oints) are added. For example, in the stem_social_psych dataset, the two variables 'belong_belongStem_w1' and 'belong_belongStem_w2' share the variable stem 'belong_belongStem' (e.g., "I feel like an outsider in STEM"), with suffixes (_w1, _w2) indicating different measurement waves. By default, missing data are excluded from the calculations in a listwise fashion.

Usage

mean_tbl(
  data,
  var_stem,
  escape_stem = FALSE,
  ignore_stem_case = FALSE,
  na_removal = "listwise",
  only = NULL,
  var_labels = NULL,
  ignore = NULL
)

Arguments

data

A data frame.

var_stem

A character string of a variable stem or the full name of a variable in data.

escape_stem

A logical value indicating whether to escape var_stem. Default is FALSE.

ignore_stem_case

A logical value indicating whether the search for columns matching the supplied var_stem is case-insensitive. Default is FALSE.

na_removal

A character string specifying how to remove missing values. Should be one of pairwise or listwise. Default is listwise.

only

A character string or vector of character strings of the kinds of summary statistics to return. Default is NULL, which returns mean (mean), standard deviation (sd), minimum value (min), maximum value (max), and non-missing responses (nobs).

var_labels

An optional named character vector or list where each element maps labels to variable names. If any element is unnamed or if any labels do not match variables in returned from data, all labels will be ignored and the table will be printed without them.

ignore

An optional vector that contains values to exclude from the data. Default is NULL, which includes all present values.

Value

A tibble presenting summary statistics for series of continuous variables with the same variable stem.

Author

Ama Nyame-Mensah

Examples

mean_tbl(data = social_psy_data,
         var_stem = "belong")
#> # A tibble: 3 × 6
#>   variable  mean    sd   min   max  nobs
#>   <chr>    <dbl> <dbl> <dbl> <dbl> <int>
#> 1 belong_1  3.83 1.10      1     5 10105
#> 2 belong_2  2.52 1.22      1     5 10105
#> 3 belong_3  3.77 0.995     1     5 10105

mean_tbl(data = social_psy_data,
         var_stem = "belong",
         na_removal = "pairwise",
         var_labels = c(belong_1 = "I feel like I belong at this institution",
                        belong_2 = "I feel like part of the community",
                        belong_3 = "I feel valued by this institution"))
#> # A tibble: 3 × 7
#>   variable variable_label                           mean    sd   min   max  nobs
#>   <chr>    <chr>                                   <dbl> <dbl> <dbl> <dbl> <int>
#> 1 belong_1 I feel like I belong at this instituti…  3.83 1.10      1     5 10138
#> 2 belong_2 I feel like part of the community        2.52 1.22      1     5 10145
#> 3 belong_3 I feel valued by this institution        3.77 0.995     1     5 10142